Conflicts:
	MyZooPets/Assets/Scenes/ZoneBedroom.unity
	MyZooPets/Assets/Scenes/ZoneYard.unity
	MyZooPets/Assets/Scripts/Minipet/Controller/MiniPetManager.cs
	MyZooPets/Assets/Scripts/Minipet/Model/MutableDataMiniPets.cs
	MyZooPets/Assets/Scripts/Minipet/SpecialMiniPet/MiniPetGameMaster.cs
	MyZooPets/Assets/Scripts/Minipet/SpecialMiniPet/MiniPetMerchant.cs
	MyZooPets/Assets/Scripts/Minipet/View/MiniPetHUDUIManager.cs

+++ MyZooPets/Assets/Scripts/Accessory/AccessoryUIManager.cs
@@ -173,7 +173,7 @@ public class AccessoryUIManager : SingletonUI<AccessoryUIManager>{
 				buyButton.isEnabled = false;
 				
 				InventoryLogic.Instance.AddItem(itemID, 1);
-				StatsController.Instance.ChangeStats(deltaStars: (int)itemData.Cost * -1);
+				StatsController.Instance.ChangeStats(deltaStars: itemData.Cost * -1);
 
 				// Change the state of the button
 				button.transform.parent.gameObject.GetComponent<AccessoryEntryUIController>().SetState(AccessoryButtonType.BoughtEquipped);
