+++ MyZooPets/Assets/Scripts/Model/ImmutableData/Item.cs
@@ -11,7 +11,7 @@ public class Item{
 	private string name; //name of item
 	private string textureName; //name of texture in the atlas
 	private ItemType type;
-	private double cost; //cost of item
+	private int cost; //cost of item
 	private CurrencyTypes currencyType;
 	protected string description;
 	private int unlockAtLevel = 0; //the level when item is unlocked
@@ -39,7 +39,7 @@ public class Item{
 		get{ return type;}
 	}
 
-	public double Cost{
+	public int Cost{
 		get{ return cost;}
 	}
 
@@ -92,13 +92,11 @@ public class Item{
 		textureName = XMLUtils.GetString(hashItemData["TextureName"] as IXMLNode);
 
 		Hashtable costAttributes = XMLUtils.GetAttributes(hashItemData["Cost"] as IXMLNode);
-
-//		cost = XMLUtils.GetInt(hashItemData["Cost"] as IXMLNode);
-
+		
 		currencyType = (CurrencyTypes)Enum.Parse(typeof(CurrencyTypes), 
 		                  HashUtils.GetHashValue<string>(costAttributes, "CurrencyType", ""));
 
-		cost = (double)Double.Parse(HashUtils.GetHashValue<string>(costAttributes, "Amount", ""));
+		cost = int.Parse(HashUtils.GetHashValue<string>(costAttributes, "Amount", ""));
 
 		description = XMLUtils.GetString(hashItemData["Desc"] as IXMLNode, "");
